+++ src/settings/components/ExternalDataSourcesConfig/ExternalDataSourcesSettings/ExternalDataSourcesSettings.js
@@ -3,13 +3,6 @@ import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 import arrayMutators from 'final-form-arrays';
 
-import { useStripes } from '@folio/stripes/core';
-import {
-  Button,
-  Pane,
-  PaneHeader,
-} from '@folio/stripes/components';
-
 import ExternalDataSourcesView from '../ExternalDataSourcesView/ExternalDataSourcesView';
 import ExternalDataSourcesLookup from '../ExternalDataSourcesLookup/ExternalDataSourcesLookup';
 import ExternalDataSourcesFormModal from '../ExternaldataSourcesFormModal/ExternalDataSourcesFormModal';
@@ -27,60 +20,30 @@ const ExternalDataSourcesSettings = ({
   onSubmit,
   initialValues
 }) => {
-  const stripes = useStripes();
-  const perm = stripes.hasPerm('ui-local-kb-admin.kbs.manage');
-  const count = externalKbs?.length ?? 0;
-
   const [mode, setMode] = useState(VIEWING);
   const [externalDataSource, setExternalDataSource] = useState();
 
-  const renderSettingsHeader = renderProps => (
-    <PaneHeader
-      {...renderProps}
-      lastMenu={perm ?
-        <Button
-          buttonStyle="primary"
-          data-test-external-data-source-new
-          id="clickable-new-external-datasource"
-          marginBottom0
-          onClick={() => setMode(CREATING)}
-        >
-          <FormattedMessage id="ui-local-kb-admin.job.new" />
-        </Button> : null
-      }
-      paneSub={<FormattedMessage id="ui-local-kb-admin.settings.externalDataSources.sourceCount" values={{ count }} />}
-      paneTitle={<FormattedMessage id="ui-local-kb-admin.section.externalDataSources" />}
-    />
-  );
-
-
   return (
     <>
-      <Pane
-        data-test-external-data-sources
-        defaultWidth="fill"
-        id="settings-external-data-sources"
-        renderHeader={renderSettingsHeader}
-      >
-        <ExternalDataSourcesLookup
-          externalKbs={externalKbs}
-          onSelectedExternalDataSource={(_e, eds) => {
-            setExternalDataSource(eds)
-          }}
-        />
-      </Pane>
+      <ExternalDataSourcesLookup
+        externalKbs={externalKbs}
+        onClick={() => setMode(CREATING)}
+        onSelectedExternalDataSource={(_e, eds) => {
+          setExternalDataSource(eds);
+        }}
+      />
       {
         externalDataSource &&
           <ExternalDataSourcesView
-              externalDataSourceId={externalDataSource.id}
-          onClose={() => setExternalDataSource()}
-          externalKbs={externalKbs}
-          initialValues={initialValues}
-          onCancel={onCancel}
-          onDelete={onDelete}
-          onSave={onSave}
-          onSubmit={onSubmit}
-          onClick={() => setMode(EDITING)}
+            externalDataSourceId={externalDataSource.id}
+            externalKbs={externalKbs}
+            initialValues={initialValues}
+            onCancel={onCancel}
+            onClick={() => setMode(EDITING)}
+            onClose={() => setExternalDataSource()}
+            onDelete={onDelete}
+            onSave={onSave}
+            onSubmit={onSubmit}
           />
       }
       <ExternalDataSourcesFormModal
