+++ src/settings/components/ExternalDataSourcesConfig/ExternalDataSourcesView/ExternalDataSourcesView.js
@@ -1,9 +1,11 @@
 import { useState } from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
+import { useQuery } from 'react-query';
 import moment from 'moment';
 
-import { useQuery } from 'react-query';
+import { useOkapiKy, useStripes } from '@folio/stripes/core';
+import { CustomMetaSection } from '@folio/stripes-erm-components';
 import {
   Col,
   KeyValue,
@@ -17,8 +19,7 @@ import {
   Icon,
   ConfirmationModal
 } from '@folio/stripes/components';
-import { useOkapiKy, useStripes} from '@folio/stripes/core';
-import { CustomMetaSection } from '@folio/stripes-erm-components';
+
 import useDisplayMetaInfo from '../useDisplayMetaInfo';
 import { KB_ENDPOINT } from '../../../../constants/endpoints';
 
@@ -81,7 +82,7 @@ const ExternalDataSourcesView = ({
     );
   };
 
-    const getActionMenu = ({ onToggle }) => {
+  const getActionMenu = ({ onToggle }) => {
     const actionsArray = [];
     if (perm) {
       actionsArray.push(
@@ -141,8 +142,7 @@ const ExternalDataSourcesView = ({
     }
 
     return (actionsArray?.length ? actionsArray : null);
-    };
-  
+  };
   const renderViewHeader = renderProps => (
     <PaneHeader
       {...renderProps}
@@ -156,94 +156,106 @@ const ExternalDataSourcesView = ({
   return (
     <>
       <Pane
-      defaultWidth="fill"
-      id="settings-externalDataSources-viewPane"
-      renderHeader={renderViewHeader}
-    >
-      {showConfirmResetSyncStatus && renderModal()}
-      <CustomMetaSection accordionLabel={syncStatus}>
-        {cursor}
-        {lastChecked}
-      </CustomMetaSection>
-      <Row>
-        <Col xs={3}>
-          <KeyValue
-            data-test-external-data-source-name
-            label={<FormattedMessage id="ui-local-kb-admin.settings.externalDataSources.name" />}
-            value={externalDataSource?.name} />
-        </Col>
-        <Col xs={5}>
-          <KeyValue
-            data-test-external-data-source-type
-            label={<FormattedMessage id="ui-local-kb-admin.settings.externalDataSources.type" />}
-            value={externalDataSource?.type} />
-        </Col>
-        <Col xs={4}>
-          <KeyValue
-            data-test-external-data-source-recordtype
-            label={<FormattedMessage id="ui-local-kb-admin.settings.externalDataSources.recordType" />}
-            value={externalDataSource?.rectype === 1 ? <FormattedMessage id="ui-local-kb-admin.settings.externalDataSources.package" /> : ''} />
-        </Col>
-      </Row>
-      <Row>
-        <Col xs={8}>
-          <KeyValue
-            data-test-external-data-source-uri
-            label={<FormattedMessage id="ui-local-kb-admin.settings.externalDataSources.uri" />}
-            value={externalDataSource?.uri ?? <NoValue />} />
-        </Col>
-        <Col xs={4}>
-          <KeyValue
-            data-test-external-data-source-trusted-source-ti
-            label={<FormattedMessage id="ui-local-kb-admin.settings.externalDataSources.trustedSourceTI" />}
-            value={<FormattedMessage id={externalDataSource?.trustedSourceTI ? 'ui-local-kb-admin.yes' : 'ui-local-kb-admin.no'} />} />
-        </Col>
-      </Row>
-      <Row>
-        <Col xs={3}>
-          <KeyValue
-            data-test-external-data-source-isactive
-            label={<FormattedMessage id="ui-local-kb-admin.settings.externalDataSources.isActive" />}
-            value={<FormattedMessage id={externalDataSource?.active ? 'ui-local-kb-admin.yes' : 'ui-local-kb-admin.no'} />} />
-        </Col>
-        <Col xs={5}>
-          <KeyValue
-            data-test-external-data-source-supportsharvesting
-            label={<FormattedMessage id="ui-local-kb-admin.settings.externalDataSources.supportsHarvesting" />}
-            value={<FormattedMessage id={externalDataSource?.supportsHarvesting ? 'ui-local-kb-admin.yes' : 'ui-local-kb-admin.no'} />} />
-        </Col>
-        <Col xs={4}>
-          <KeyValue
-            data-test-external-data-source-activationenabled
-            label={<FormattedMessage id="ui-local-kb-admin.settings.externalDataSources.activationEnabled" />}
-            value={<FormattedMessage id={externalDataSource?.activationEnabled ? 'ui-local-kb-admin.yes' : 'ui-local-kb-admin.no'} />} />
-        </Col>
-      </Row>
-      <Row>
-        <Col xs={3}>
-          <KeyValue
-            data-test-external-data-source-listprefix
-            label={<FormattedMessage id="ui-local-kb-admin.settings.externalDataSources.listPrefix" />}
-            value={externalDataSource?.listPrefix ?? <NoValue />} />
-        </Col>
-        <Col xs={5}>
-          <KeyValue
-            data-test-external-data-source-fullprefix
-            label={<FormattedMessage id="ui-local-kb-admin.settings.externalDataSources.fullPrefix" />}
-            value={externalDataSource?.fullPrefix ?? <NoValue />} />
-        </Col>
-        <Col xs={4}>
-          <KeyValue
-            data-test-external-data-source-principal
-            label={<FormattedMessage id="ui-local-kb-admin.settings.externalDataSources.principal" />}
-            value={externalDataSource?.principal ?? <NoValue />} />
-        </Col>
-      </Row>
-      <KeyValue
-        data-test-external-data-source-credentials
-        label={<FormattedMessage id="ui-local-kb-admin.settings.externalDataSources.credentials" />}
-        value={externalDataSource?.credentials ?? <NoValue />} />
-    </Pane>
+        defaultWidth="fill"
+        id="settings-externalDataSources-viewPane"
+        renderHeader={renderViewHeader}
+      >
+        {showConfirmResetSyncStatus && renderModal()}
+        <CustomMetaSection accordionLabel={syncStatus}>
+          {cursor}
+          {lastChecked}
+        </CustomMetaSection>
+        <Row>
+          <Col xs={3}>
+            <KeyValue
+              data-test-external-data-source-name
+              label={<FormattedMessage id="ui-local-kb-admin.settings.externalDataSources.name" />}
+              value={externalDataSource?.name}
+            />
+          </Col>
+          <Col xs={5}>
+            <KeyValue
+              data-test-external-data-source-type
+              label={<FormattedMessage id="ui-local-kb-admin.settings.externalDataSources.type" />}
+              value={externalDataSource?.type}
+            />
+          </Col>
+          <Col xs={4}>
+            <KeyValue
+              data-test-external-data-source-recordtype
+              label={<FormattedMessage id="ui-local-kb-admin.settings.externalDataSources.recordType" />}
+              value={externalDataSource?.rectype === 1 ? <FormattedMessage id="ui-local-kb-admin.settings.externalDataSources.package" /> : ''}
+            />
+          </Col>
+        </Row>
+        <Row>
+          <Col xs={8}>
+            <KeyValue
+              data-test-external-data-source-uri
+              label={<FormattedMessage id="ui-local-kb-admin.settings.externalDataSources.uri" />}
+              value={externalDataSource?.uri ?? <NoValue />}
+            />
+          </Col>
+          <Col xs={4}>
+            <KeyValue
+              data-test-external-data-source-trusted-source-ti
+              label={<FormattedMessage id="ui-local-kb-admin.settings.externalDataSources.trustedSourceTI" />}
+              value={<FormattedMessage id={externalDataSource?.trustedSourceTI ? 'ui-local-kb-admin.yes' : 'ui-local-kb-admin.no'} />}
+            />
+          </Col>
+        </Row>
+        <Row>
+          <Col xs={3}>
+            <KeyValue
+              data-test-external-data-source-isactive
+              label={<FormattedMessage id="ui-local-kb-admin.settings.externalDataSources.isActive" />}
+              value={<FormattedMessage id={externalDataSource?.active ? 'ui-local-kb-admin.yes' : 'ui-local-kb-admin.no'} />}
+            />
+          </Col>
+          <Col xs={5}>
+            <KeyValue
+              data-test-external-data-source-supportsharvesting
+              label={<FormattedMessage id="ui-local-kb-admin.settings.externalDataSources.supportsHarvesting" />}
+              value={<FormattedMessage id={externalDataSource?.supportsHarvesting ? 'ui-local-kb-admin.yes' : 'ui-local-kb-admin.no'} />}
+            />
+          </Col>
+          <Col xs={4}>
+            <KeyValue
+              data-test-external-data-source-activationenabled
+              label={<FormattedMessage id="ui-local-kb-admin.settings.externalDataSources.activationEnabled" />}
+              value={<FormattedMessage id={externalDataSource?.activationEnabled ? 'ui-local-kb-admin.yes' : 'ui-local-kb-admin.no'} />}
+            />
+          </Col>
+        </Row>
+        <Row>
+          <Col xs={3}>
+            <KeyValue
+              data-test-external-data-source-listprefix
+              label={<FormattedMessage id="ui-local-kb-admin.settings.externalDataSources.listPrefix" />}
+              value={externalDataSource?.listPrefix ?? <NoValue />}
+            />
+          </Col>
+          <Col xs={5}>
+            <KeyValue
+              data-test-external-data-source-fullprefix
+              label={<FormattedMessage id="ui-local-kb-admin.settings.externalDataSources.fullPrefix" />}
+              value={externalDataSource?.fullPrefix ?? <NoValue />}
+            />
+          </Col>
+          <Col xs={4}>
+            <KeyValue
+              data-test-external-data-source-principal
+              label={<FormattedMessage id="ui-local-kb-admin.settings.externalDataSources.principal" />}
+              value={externalDataSource?.principal ?? <NoValue />}
+            />
+          </Col>
+        </Row>
+        <KeyValue
+          data-test-external-data-source-credentials
+          label={<FormattedMessage id="ui-local-kb-admin.settings.externalDataSources.credentials" />}
+          value={externalDataSource?.credentials ?? <NoValue />}
+        />
+      </Pane>
       {deleteModal && (
         <ConfirmationModal
           buttonStyle="danger"
